// src/api/generated/model/perioExamDetailRead.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * Reckon Dental Backend
 * Dental PMS REST API. Column-tenant-scoped; snake_case; Orval-ready.
 * OpenAPI spec version: 1.0.0
 */

export interface PerioExamDetailRead {
  id: number;
  exam_id: number;
  tooth_no: string;
  created_at: string;
  updated_at?: string | null;
  created_by?: number | null;
  updated_by?: number | null;
  created_by_name?: string | null;
  updated_by_name?: string | null;
  pd1?: number | null;
  fgm1?: number | null;
  mgj1?: number | null;
  cal1?: number | null;
  furc1?: number | null;
  bleed1?: boolean | null;
  supp1?: boolean | null;
  pd2?: number | null;
  fgm2?: number | null;
  mgj2?: number | null;
  cal2?: number | null;
  furc2?: number | null;
  bleed2?: boolean | null;
  supp2?: boolean | null;
  pd3?: number | null;
  fgm3?: number | null;
  mgj3?: number | null;
  cal3?: number | null;
  furc3?: number | null;
  bleed3?: boolean | null;
  supp3?: boolean | null;
  pd4?: number | null;
  fgm4?: number | null;
  mgj4?: number | null;
  cal4?: number | null;
  furc4?: number | null;
  bleed4?: boolean | null;
  supp4?: boolean | null;
  pd5?: number | null;
  fgm5?: number | null;
  mgj5?: number | null;
  cal5?: number | null;
  furc5?: number | null;
  bleed5?: boolean | null;
  supp5?: boolean | null;
  pd6?: number | null;
  fgm6?: number | null;
  mgj6?: number | null;
  cal6?: number | null;
  furc6?: number | null;
  bleed6?: boolean | null;
  supp6?: boolean | null;
  mobility_buccal?: string | null;
  mobility_lingual?: string | null;
}